test: Add WindowTest
