+++ UnitTests/Utils Test.vcxproj
@@ -116,6 +116,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="UtilsTest.cpp" />
+    <ClCompile Include="WindowTest.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\VimHotkeyWindowManager\VimHotkeyWindowManager.vcxproj">
