+++ VimHotkeyWindowManager/window.cpp
@@ -29,18 +29,3 @@ Window* FindWindowByHwnd(HWND hwnd) {
     return nullptr; 
 }
 
-// Find the monitor in the global list by the HMONITOR
-// If the monitor is found, return the pointer to the monitor
-// Otherwise, return nullptr
-Monitor *FindMonitorByHmonitor(HMONITOR hMonitor) {
-	Monitor* current = primary_monitor;
-	while (current) {
-		if (current->hMonitor == hMonitor) {
-			return current;
-		}
-		current = current->next_monitor;
-	}
-
-	return nullptr;
-}
-
